change the name of sqlplus
since sqlplus is an oracle product, to avoid confusion we are
changing the name of the camel component to sqletl.  We will retain
the old name for the sake of backwards compatibility.  May deprecate
it in a later release.

Also updated a couple of tests to use the non-deprecated version of the
camel api and upgraded the version of camelGLite that we are using