update MDOC_VERSION
for some reason the task to update MDOC_VERSION no longer exists.  I
have added this task back so the command line utility will properly
print out the version of the command line utility.  So now,
whenever the project is compiled, the version file in
metridoc-job-cli is updated from the root version file.