fix import jobs bug
when an optional subdirectory is provided, the import jobs command
was not working properly.  It would try to import jobs according to
the zip file instead of a sub directory of the zip file.  In the
process of doing this it would also create another directory for a
job with nothing in it.