add codenarc
added the rules file from the lazybones project along with codenarc
configuration for all subprojects.  metridoc-job-cli needs to have a
couple of codenarc tasks turned off to avoid issues with tests.
these tasks don't seem to be called in the other projects.  Once we
get more familiar with this tool, maybe we will figure out what is
going on.  For now, it is there.